LayoutTests/media/track/track-webvtt-tc025-class-markup-expected.txt

Tests cues with class markup <c>.

EXPECTED (video.textTracks[0].cues.length == '3') OK
EXPECTED (true == 'true') OK
EXPECTED (true == 'true') OK
EXPECTED (true == 'true') OK
EXPECTED (video.textTracks[1].cues.length == '3') OK
EXPECTED (true == 'true') OK
EXPECTED (true == 'true') OK
EXPECTED (true == 'true') OK
END OF TEST
